cashflow_project/requirements.txt
Django==5.0.3
django-filter==24.2